+++ src/remark-calc-to-code.js
@@ -1,24 +1,18 @@
-import { visit } from 'unist-util-visit'
+// import { visit } from "unist-util-visit";
 
-export const remarkMath2Code = () => {
-  return tree => {
-    visit(tree, { type: 'calc' }, element => {
-      element.type = 'code'
-      element.lang = 'calc'
-      element.data.hChildren = undefined
-      element.data.hName = undefined
-      element.data.hProperties = {
-        lang: 'calc'
-      }
-    })
-    visit(tree, { type: 'inlineCalc' }, element => {
-      element.type = 'inlineCode'
-      element.lang = 'inline_calc'
-      element.data.hChildren = undefined
-      element.data.hName = undefined
-      element.data.hProperties = {
-        lang: 'inline_calc'
-      }
-    })
-  }
-}
+// here you need to create a parser for inline calc
+
+// export const remarkCalc2Code = () => {
+//   return tree => {
+//     visit(tree, { type: 'inlineCalc' }, element => {
+//       console.log("inlineCalc element:", element)
+//       element.type = 'inlineCode'
+//       element.lang = 'inline_calc'
+//       element.data.hChildren = undefined
+//       element.data.hName = undefined
+//       element.data.hProperties = {
+//         lang: 'inline_calc'
+//       }
+//     })
+//   }
+// }