+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.0 - First Release
-* Basic inline calculator published and tested(kind of)
+
+- Basic inline calculator published and tested(kind of)
 
 ## 0.2.0
-* Show positives now on by default
-* Can be turned on or off via plugin menus
+
+- Show positives now on by default
+- Can be turned on or off via plugin menus